Widen 23 of 24 narrow return types; only one genuinely needs to be narrow
Following the func_80059FAC lesson to its conclusion. Every u8/u16/s8/s16
return in src/ is an andi waiting for its first caller, and the callee stays
green either way, so the mistake surfaces as a caller problem long after it
was made. Widened all 24 to s32 and rebuilt: 23 are byte-identical either way.

The one that is not is func_800451E0, which returns `s16` and comes out one
instruction shorter as s32 -- there the narrowing is doing real work (a sign
extension retail performs), so it stays.

Left func_80047008 out of the first pass by mistake -- it was in the
cascade-fallout list rather than the genuinely-broken one, and my revert
could not tell those apart. Verified separately: it widens cleanly too.

+++ src/func_80021558.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-s16 func_80021558(s32 arg0, s32 arg1) {
+s32 func_80021558(s32 arg0, s32 arg1) {
     Threshold *p = D_801798A8[arg0];
 
     while (1) {
