+++ src/func_80024060.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-u8 func_80024060(u8 *arg0) {
+s32 func_80024060(u8 *arg0) {
     func_80023FBC(arg0);
 
     return arg0[0x19];
