+++ src/func_8003F70C.c
@@ -1,6 +1,6 @@
 #include "common.h"
 
-u8 func_8003F70C(void) {
+s32 func_8003F70C(void) {
     func_8003F454();
 
     if (D_8009B3FA != 0) {
