+++ src/func_80044FE4.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-s16 func_80044FE4(void) {
+s32 func_80044FE4(void) {
     return *(s16 *)(D_8009B45C + 0x510);
 }
