+++ src/func_80047008.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_80047008(void) {
+s32 func_80047008(void) {
     return D_8009B45C[0x48];
 }
