+++ src/func_80058DC0.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_80058DC0(void) {
+s32 func_80058DC0(void) {
     return D_8009AFA6;
 }
