+++ src/func_800744F4.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u16 func_800744F4(void) {
+s32 func_800744F4(void) {
     return *D_80092A90;
 }
