+++ src/func_8007BF50.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8007BF50(void) {
+s32 func_8007BF50(void) {
     return D_80093898;
 }
