+++ src/func_8007F5C4.c
@@ -1,5 +1,5 @@
 #include "common.h"
 
-u8 func_8007F5C4(void) {
+s32 func_8007F5C4(void) {
     return D_8009466A;
 }
